+++ Hometask with CatalogTree/Code/catalog.h
@@ -5,13 +5,13 @@
 
 NODE* MakeTree(char* path, NODE* root);
 
-Node* Find(char* path, FILE* fout, int* size, Node* root);
+Node* Find(char* path, FILE* fout, unsigned long int* size, Node* root);
 
 void FindCatalogs(NODE* root, int level, FILE* fout, char* target_file);
 
 void NumberOfFilesInDirectory(NODE* root, int* number);
 
-void FindTheHugeFiles(NODE* root, FILE* fout, char* path, int* size);
+void FindTheHugeFiles(NODE* root, FILE* fout, char* path, unsigned long int* size);
 
 #endif // !CATALOG_H
